Move Object definition to its own file
